+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Upcoming
-* Added `World`>`Weather Feature`
+* Added `World`>`Weather` Feature
   * Added Thunderstorm intensity, making thunderstorms have more chance to lightning strike when intensity is higher.
     Intensity varies every few minutes
 * Entities now have 20% chance to fall from their mount